rename symbols_actor.py to symbols_actor_base.py
