+++ pulse/interface/viewer_3d/actors/symbols_actor_base.py
